+++ docs/changelog.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 2.0.0
+
+* Revise access permissions for some views:
+
+  * ImageSet detail view now accessible by any authenticated user
+  * Image delete view now accessible only by image owner.
+  * Image "toggle primary" view now accessible only by image owner.
+
+## 1.0.0
+
+* Update version for Pinax 16.04 release
+
 ## 0.2.1
 
 * Improve documentation